backend/simple_worker/requirements.txt
celery[redis]
pymysql
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
requests
minio
gevent
pytz